separate displayHint into presentation and container component
